create_project_files_realtime.py
# ==============================================================================
# ファイル: create_project_files_realtime.py
# 説明: このスクリプトは、リアルタイム自動トレードシステムに
#       必要なファイルとディレクトリの骨格を生成します。
# バージョン: v1.0
# ==============================================================================
import os

project_files_realtime = {
    "config_realtrade.py": """import os

# --- API認証情報 ---
# セキュリティのため、環境変数から読み込むことを強く推奨します。
# 例: export API_KEY='your_key'
API_KEY = os.getenv("API_KEY", "YOUR_API_KEY_HERE")
API_SECRET = os.getenv("API_SECRET", "YOUR_API_SECRET_HERE")

# --- 銘柄・戦略設定 ---
# 最新のall_recommend_*.csvを自動で検索するためのパターン
RECOMMEND_FILE_PATTERN = "all_recommend_*.csv"

# --- 安全装置設定 ---
# 1注文あたりの最大金額(円)。これを超える注文はブロックされる。
MAX_ORDER_SIZE_JPY = 1000000
# 同時に発注できる最大注文数
MAX_CONCURRENT_ORDERS = 5
# 資産がこの割合(マイナス値)以上に減少したらシステムを緊急停止する
EMERGENCY_STOP_THRESHOLD = -0.1

# --- データベース設定 ---
BASE_DIR = os.path.dirname(os.path.abspath(__file__))
# realtrade/db ディレクトリ内にデータベースファイルを配置
DB_PATH = os.path.join(BASE_DIR, "realtrade", "db", "realtrade_state.db")
""",

    "run_realtrade.py": """import logging
import schedule
import time
from datetime import datetime
import backtrader as bt
import yaml
import pandas as pd
import glob
import os

# import config_realtrade as config
# import logger_setup
# import btrader_strategy
# from realtrade.state_manager import StateManager
# from realtrade.broker_bridge import SbiBrokerBridge # 実装例
# from realtrade.data_fetcher import SbiDataFetcher # 実装例

# logger_setup.setup_logging()
# logger = logging.getLogger(__name__)

class RealtimeTrader:
    def __init__(self):
        print("リアルタイムトレーダーを初期化中...")
        # logger.info("リアルタイムトレーダーを初期化中...")
        # self.strategy_catalog = self._load_strategy_catalog('strategies.yml')
        # self.strategy_assignments = self._load_strategy_assignments(config.RECOMMEND_FILE_PATTERN)
        # self.state_manager = StateManager(config.DB_PATH)
        # self.broker_bridge = SbiBrokerBridge() # 後続ステップで実装
        # self.data_fetcher = SbiDataFetcher()   # 後続ステップで実装
        # self.cerebro = self._setup_cerebro()
        self.is_running = False

    def _load_strategy_catalog(self, filepath):
        # logger.info(f"戦略カタログをロード: {filepath}")
        with open(filepath, 'r', encoding='utf-8') as f:
            strategies = yaml.safe_load(f)
        return {s['name']: s for s in strategies}

    def _load_strategy_assignments(self, filepath_pattern):
        # logger.info(f"銘柄・戦略対応ファイルを検索: {filepath_pattern}")
        files = glob.glob(filepath_pattern)
        if not files:
            # logger.error("銘柄・戦略対応ファイルが見つかりません。")
            raise FileNotFoundError(f"銘柄・戦略対応ファイルが見つかりません: {filepath_pattern}")
        latest_file = max(files, key=os.path.getctime)
        # logger.info(f"最新の対応ファイルをロード: {latest_file}")
        df = pd.read_csv(latest_file)
        return pd.Series(df.strategy_name.values, index=df.symbol).to_dict()

    def _setup_cerebro(self):
        # logger.info("Cerebroエンジンをセットアップ中...")
        # cerebro = bt.Cerebro(runonce=False) # リアルタイムなのでrunonce=False
        # # ... Broker, Data, Strategy の設定は後続ステップで実装 ...
        # return cerebro
        pass

    def run_job(self):
        print(f"{datetime.now()}: 取引ジョブを実行中...")
        # logger.info("取引ジョブを実行中...")
        # self.cerebro.run()
        pass

    def start(self):
        print("システムを開始します。取引時間まで待機...")
        # logger.info("システムを開始します。取引時間まで待機...")
        # schedule.every().day.at("08:55").do(self.run_job) # 寄り付き前に起動
        # schedule.every().day.at("15:05").do(self.stop)  # 大引け後に停止
        self.is_running = True
        self.run_job() # テストのため即時実行

    def stop(self):
        print("システムを停止します。")
        # logger.info("システムを停止します。")
        # self.state_manager.close()
        # ... 注文キャンセルやポジション保存処理 ...
        self.is_running = False
        print("システムが正常に停止しました。")

if __name__ == '__main__':
    trader = RealtimeTrader()
    try:
        trader.start()
        while trader.is_running:
            # schedule.run_pending()
            time.sleep(1)
    except KeyboardInterrupt:
        print("Ctrl+Cを検知しました。システムを安全に停止します...")
        trader.stop()
    except Exception as e:
        print(f"予期せぬエラーが発生しました: {e}")
        # logger.critical(f"予期せぬエラーでシステムが停止しました。", exc_info=True)
        trader.stop()
""",

    "realtrade/__init__.py": """# このファイルは'realtrade'ディレクトリをPythonパッケージとして認識させるためのものです。
""",

    "realtrade/broker_bridge.py": """import backtrader as bt
import abc

class BrokerBridge(bt.broker.BrokerBase, metaclass=abc.ABCMeta):
    \"\"\"
    証券会社APIと連携するための抽象基底クラス。
    backtraderのBrokerBaseを継承します。
    \"\"\"
    @abc.abstractmethod
    def get_cash(self):
        \"\"\"利用可能な現金額を取得します。\"\"\"
        raise NotImplementedError

    @abc.abstractmethod
    def get_position(self, data, clone=True):
        \"\"\"指定された銘柄のポジションを取得します。\"\"\"
        raise NotImplementedError

    # place_order, cancel_order などのメソッドも後続ステップで定義
""",

    "realtrade/data_fetcher.py": """import backtrader as bt
import abc

class DataFetcher(metaclass=abc.ABCMeta):
    \"\"\"
    証券会社APIから価格データを取得するための抽象基底クラス。
    \"\"\"
    @abc.abstractmethod
    def start(self):
        \"\"\"データ取得を開始します (WebSocket接続など)。\"\"\"
        raise NotImplementedError

    @abc.abstractmethod
    def stop(self):
        \"\"\"データ取得を停止します。\"\"\"
        raise NotImplementedError

    @abc.abstractmethod
    def get_data_feed(self, symbol):
        \"\"\"指定された銘柄のデータフィードオブジェクトを返します。\"\"\"
        raise NotImplementedError

    @abc.abstractmethod
    def fetch_historical_data(self, symbol, period, timeframe):
        \"\"\"戦略の初期化に必要な過去のデータを取得します。\"\"\"
        raise NotImplementedError
""",

    "realtime/state_manager.py": """import sqlite3
import logging
import os

# logger = logging.getLogger(__name__)

class StateManager:
    def __init__(self, db_path):
        \"\"\"
        システムの稼働状態を永続化・復元するためのクラス。
        :param db_path: SQLiteデータベースファイルのパス。
        \"\"\"
        self.db_path = db_path
        self.conn = None
        # DBファイルが置かれるディレクトリがなければ作成
        os.makedirs(os.path.dirname(db_path), exist_ok=True)
        try:
            self.conn = sqlite3.connect(db_path, check_same_thread=False)
            self._create_tables()
            # logger.info(f"データベースに接続しました: {db_path}")
            print(f"データベースに接続しました: {db_path}")
        except sqlite3.Error as e:
            # logger.error(f"データベース接続エラー: {e}")
            print(f"データベース接続エラー: {e}")
            raise

    def _create_tables(self):
        \"\"\"必要なテーブルが存在しない場合に作成する。\"\"\"
        cursor = self.conn.cursor()
        # ポジション管理テーブル
        cursor.execute('''
            CREATE TABLE IF NOT EXISTS positions (
                symbol TEXT PRIMARY KEY,
                size REAL NOT NULL,
                price REAL NOT NULL,
                entry_datetime TEXT NOT NULL
            )
        ''')
        # 注文管理テーブル
        cursor.execute('''
            CREATE TABLE IF NOT EXISTS orders (
                order_id TEXT PRIMARY KEY,
                symbol TEXT NOT NULL,
                order_type TEXT NOT NULL,
                size REAL NOT NULL,
                price REAL,
                status TEXT NOT NULL
            )
        ''')
        self.conn.commit()
        # logger.info("テーブルの初期化を確認しました。")
        print("データベーステーブルの初期化を確認しました。")

    def close(self):
        if self.conn:
            self.conn.close()
            # logger.info("データベース接続をクローズしました。")
            print("データベース接続をクローズしました。")
"""
}

def create_files(files_dict):
    """
    指定された辞書に基づいてプロジェクトファイルとディレクトリを生成します。
    """
    for filename, content in files_dict.items():
        # ディレクトリが存在しない場合は作成
        if os.path.dirname(filename) and not os.path.exists(os.path.dirname(filename)):
            os.makedirs(os.path.dirname(filename))

        content = content.strip()
        try:
            with open(filename, 'w', encoding='utf-8', newline='\n') as f:
                f.write(content)
            print(f"ファイルを作成/更新しました: {filename}")
        except IOError as e:
            print(f"エラー: ファイル '{filename}' の作成に失敗しました。 - {e}")

if __name__ == '__main__':
    print("リアルタイムトレード用のプロジェクトファイル生成を開始します...")
    create_files(project_files_realtime)
    print("\nリアルタイムトレード用のプロジェクトファイルの生成が完了しました。")
    print("\n--- 実行方法 ---")
    print("1. このスクリプトを `create_project_files_realtime.py` として保存します。")
    print("2. ターミナルで `python create_project_files_realtime.py` を実行します。")

