requirements.txt
backtrader
pandas==2.1.4
numpy==1.26.4
PyYAML==6.0.1
matplotlib
plotly==5.18.0
Flask==3.0.0
schedule